removed webpack and use vite to build
