--- .babelrc
@@ -1,8 +1,0 @@
-{
-    "presets": ["@babel/preset-env", "@babel/preset-react"],
-    "env": {
-        "development": {
-            "presets": []
-        }
-    }
-}
