+++ client/components/Partners.js
@@ -1,73 +1,88 @@
+import CoinlistImg from "../assets/images/investors/Coinlist.jpg.webp";
+import RepublicImg from "../assets/images/investors/Republic.jpg.webp";
+import SuperfarmImg from "../assets/images/investors/Superfarm.jpg.webp";
+import NeoTokyoImg from "../assets/images/investors/NeoTokyo.jpg.webp";
+import UnixImg from "../assets/images/investors/Unix.jpg.webp";
+import MeritCircleImg from "../assets/images/investors/Sin.City_website_backers_merit_circle.png.webp";
+import YieldGuildImg from "../assets/images/investors/yieldguild.jpg.webp";
+import PerionDAOImg from "../assets/images/investors/PerionDAO.jpg.webp";
+import GuildFiImg from "../assets/images/investors/GuildFi.jpg.webp";
+import BinanceNFTImg from "../assets/images/investors/BinanceNFT.jpg.webp";
+import FractalImg from "../assets/images/investors/fractal.jpg.webp";
+import CoinMarketCapImg from "../assets/images/investors/CoinMarketCap.jpg.webp";
+import IlluviumImg from "../assets/images/investors/Illuvium.jpg.webp";
+import SynthetixImg from "../assets/images/investors/Synthetix.jpg.webp";
+
 const PARTNERS = [
   {
     href: "https://coinlist.co/",
     name: "Coinlist",
-    img: require("../assets/images/investors/Coinlist.jpg.webp"),
+    img: CoinlistImg,
   },
   {
     href: "https://republic.com/",
     name: "Republic",
-    img: require("../assets/images/investors/Republic.jpg.webp"),
+    img: RepublicImg,
   },
   {
     href: "https://superfarm.com/",
     name: "Superfarm",
-    img: require("../assets/images/investors/Superfarm.jpg.webp"),
+    img: SuperfarmImg,
   },
   {
     href: "https://neotokyo.codes/",
     name: "Neo Tokyo",
-    img: require("../assets/images/investors/NeoTokyo.jpg.webp"),
+    img: NeoTokyoImg,
   },
   {
     href: "https://www.unixgaming.org/",
     name: "Unix",
-    img: require("../assets/images/investors/Unix.jpg.webp"),
+    img: UnixImg,
   },
   {
     href: "https://meritcircle.io/",
     name: "Merit Circle",
-    img: require("../assets/images/investors/Sin.City_website_backers_merit_circle.png.webp"),
+    img: MeritCircleImg,
   },
   {
     href: "https://yieldguild.io/",
     name: "Yield Guild",
-    img: require("../assets/images/investors/yieldguild.jpg.webp"),
+    img: YieldGuildImg,
   },
   {
     href: "https://perion.gg/",
     name: "PerionDAO",
-    img: require("../assets/images/investors/PerionDAO.jpg.webp"),
+    img: PerionDAOImg,
   },
   {
     href: "https://guildfi.com/",
     name: "GuildFi",
-    img: require("../assets/images/investors/GuildFi.jpg.webp"),
+    img: GuildFiImg,
   },
   {
     href: "https://www.binance.com/en/nft/home",
     name: "Binance NFT",
-    img: require("../assets/images/investors/BinanceNFT.jpg.webp"),
+    img: BinanceNFTImg,
   },
   {
     href: "https://www.fractal.is/",
     name: "Fractal",
-    img: require("../assets/images/investors/fractal.jpg.webp"),
+    img: FractalImg,
   },
   {
     href: "https://coinmarketcap.com/",
     name: "Coinmarketcap",
-    img: require("../assets/images/investors/CoinMarketCap.jpg.webp"),
+    img: CoinMarketCapImg,
   },
   {
     href: "https://www.illuvium.io/",
     name: "Illuvium",
-    img: require("../assets/images/investors/Illuvium.jpg.webp"),
+    img: IlluviumImg,
   },
   {
     href: "https://synthetix.io/",
     name: "Synthetix",
-    img: require("../assets/images/investors/Synthetix.jpg.webp"),
+    img: SynthetixImg,
   },
 ];
