+++ client/utils/ClientApi.js
@@ -1,4 +1,3 @@
-import "@babel/polyfill";
 import superagent from "superagent";
 
 class ClientApi {
