+++ index.html
@@ -1,0 +1,23 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <meta name="description" content="Website for Mobland" />
+    <title>Mobland</title>
+    <link rel="icon" href="/favicon.ico" />
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+  </head>
+  <body class="d-flex flex-column h-100">
+    <script type="module" src="/client/index.jsx"></script>
+    <noscript>
+      MOBLAND is the first-ever Mafia themed Metaverse backed by super angels,
+      such as Justin Kan (co-founder of Twitch), Do Kwon (co-founder of Terra)
+      and more.<br />
+      It is developed by a team of industry experts from Disney, Ubisoft,
+      Roblox, EA, Gameloft, Google, Yahoo and Twitter.<br />
+      You need to enable JavaScript to run this app.</noscript
+    >
+    <div id="app"></div>
+  </body>
+</html>
