+++ package.json
@@ -18,6 +18,7 @@
     "@mui/system": "^5.11.7",
     "@superpowerlabs/salus": "0.1.0-beta.1",
     "body-parser": "^1.19.0",
+    "bootstrap": "^5.3.0",
     "chalk": "^2.4.2",
     "cli-spinner": "^0.2.10",
     "compression-webpack-plugin": "^10.0.0",
@@ -45,46 +46,34 @@
     "video-react": "^0.16.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.21.8",
-    "@babel/polyfill": "^7.12.1",
-    "@babel/preset-env": "^7.21.5",
-    "@babel/preset-react": "^7.18.6",
-    "@babel/register": "^7.21.0",
     "@testing-library/jest-dom": "^5.16.5",
-    "babel-eslint": "^7.2.3",
-    "babel-loader": "^9.1.2",
-    "babel-plugin-react-transform": "^2.0.2",
-    "babel-preset-stage-0": "^6.24.1",
+    "@vitejs/plugin-react": "^4.0.0",
+    "@vitejs/plugin-react-refresh": "^1.3.6",
+    "@wojtekmaj/enzyme-adapter-react-17": "^0.8.0",
     "chai": "^4.3.7",
     "css-loader": "^3.6.0",
-    "css-minimizer-webpack-plugin": "^4.2.2",
+    "enzyme": "^3.11.0",
     "eslint": "^6.8.0",
     "eslint-plugin-react": "^7.32.2",
     "glob": "^9.3.2",
-    "html-webpack-plugin": "^5.5.0",
     "husky": "^8.0.0",
     "jest": "^29.5.0",
-    "mini-css-extract-plugin": "^2.7.2",
     "mocha": "^8.3.2",
-    "node-polyfill-webpack-plugin": "^2.0.0",
     "prettier": "2.5.1",
     "pretty-quick": "^3.1.3",
-    "purgecss-webpack-plugin": "^5.0.0",
+    "react-test-renderer": "^18.2.0",
     "regenerator-runtime": "^0.13.11",
     "style-loader": "^3.3.1",
     "supertest": "^6.3.3",
-    "webpack": "^5.73.0",
-    "webpack-cli": "^4.10.0",
-    "webpack-deadcode-plugin": "^0.1.17",
-    "webpack-merge": "^5.8.0"
+    "vite": "^4.3.7",
+    "vite-plugin-compression": "^0.5.1",
+    "vite-plugin-node-polyfills": "^0.8.2"
   },
   "scripts": {
-    "start": "node .",
-    "start:dev": "npm run dev",
-    "build": "webpack --config webpack.prod.js",
+    "start": "vite",
+    "build": "vite build",
+    "serve": "vite preview",
     "format": "npx prettier --write .",
-    "dev": "webpack --config webpack.dev.js && NODE_ENV=development node .",
-    "prod": "webpack --config webpack.prod.js && NODE_ENV=development node .",
     "prepare": "husky install",
     "postinstall": "bin/post-install.sh",
     "test": "npm run test:server; npm run test:client -- --watchAll=false",
